Move kseq++ to its own subdirectory
