+++ src/readlen.hpp
@@ -1,7 +1,7 @@
 #ifndef STROBEALIGN_READLEN_HPP
 #define STROBEALIGN_READLEN_HPP
 
-#include "kseq++.hpp"
+#include "kseq++/kseq++.hpp"
 #include <zlib.h>
 #include "pc.hpp"
